# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED, INVALID_TOKEN, CONFLICT]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    FeedItem:
      type: object
      description: >
        Immutable v1 contract — exactly 11 fields.
        No renames, no removals allowed in v1. Breaking changes go to /v2/*.
      required:
        - id
        - media_preview
        - price_display
        - title
        - location
        - trust_signal
        - has_video
        - is_sponsored
      properties:
        id:
          type: string
          description: Listing identifier (use for detail navigation).
        media_preview:
          type: string
        price_display:
          type: string
        installment_badge:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        urgency_signal:
          type: string
          nullable: true
        trust_signal:
          type: string
        smart_badge:
          type: string
          nullable: true
        has_video:
          type: boolean
        is_sponsored:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        best_offer_badge:
          type: string
          nullable: true
        industrial_type:
          type: string
          nullable: true
          description: Industrial sub-type (factory, warehouse, land, production_line, raw_material, machine) used for client-side category grouping. Null for non-industrial listings.
        origin_type:
          type: string
          nullable: true
          description: '"imported" when the listing is imported (else null) — drives the "مستورد / Imported" card badge on industrial + car listings.'
        category:
          type: string
          enum: [car, real_estate, industrial]
          nullable: true
          description: Additive. The listing's main section — lets clients adapt per-section UI (e.g. the save glyph) without re-deriving it.
        is_active:
          type: boolean
          nullable: true
          description: True when the listing status is "active" (publicly visible and promotable). Lets owner-facing surfaces built on FeedItem (e.g. the profile grid) gate the Promote control without fetching ListingDetail. Null when status is unknown.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: True when the seller opted this listing in to WhatsApp contact. Opt-in only (default false). Lets surfaces decide whether to surface a WhatsApp action. Null when unknown.
        created_at:
          type: string
          nullable: true
          description: Listing creation timestamp (ISO 8601). Owner-facing grids (profile, "my listings") render a "Listed <date>" caption. Null only when the source row lacks it; public feeds always carry it.
        is_request:
          type: boolean
          nullable: true
          description: True when this is a buyer "request/wanted" post (looking to buy) rather than a sale listing. Surfaces render a "طلب / Wanted" badge and the requests filter keys off it. Null when unknown.
        is_bookable:
          type: boolean
          nullable: true
          description: >-
            Additive. True when this listing is a furnished/daily rental (hotel
            model) that can be reserved from its detail. Lets the feed, search
            card, and map pin surface a "قابل للحجز / Bookable" affordance without
            fetching the detail. Null/false for everything else.
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
    MediaItem:
      type: object
      required: [id, type, url, is_thumbnail]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [image, video]
        url:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        is_thumbnail:
          type: boolean
...
    ListingDetail:
      type: object
      required:
        - id
        - title
        - category
        - price_display
        - location
        - status
        - created_at
        - media
        - specs
        - payment
        - seller
        - interactions
        - is_saved
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [car, real_estate, industrial]
        price_display:
          type: string
        price_cash:
          type: number
          nullable: true
          description: >-
            Additive. Raw numeric cash price. For furnished/daily rentals this is
            the per-night rate the booking widget multiplies by the night count
            for a pre-booking estimate; the server stays authoritative on the
            real total. Null when unknown.
        location:
          type: string
        status:
          type: string
          enum: [active, sold, archived]
        created_at:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        specs:
          type: object
          additionalProperties: true
        payment:
          $ref: "#/components/schemas/PaymentSummary"
        seller:
          $ref: "#/components/schemas/Seller"
        interactions:
          type: object
          properties:
            views:
              type: integer
            clicks:
              type: integer
        is_saved:
          type: boolean
        is_request:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        best_offer:
          $ref: "#/components/schemas/Offer"
          nullable: true
        linked_listings:
          type: array
          description: >
            Additive (Task #33). Bidirectional supply-chain cross-links surfaced
            on the listing detail only. Optional — older clients may ignore it.
          items:
            $ref: "#/components/schemas/LinkedListing"
        logistics:
          allOf:
            - $ref: "#/components/schemas/Logistics"
          nullable: true
          description: >
            Additive (Task #40). Optional seller-provided logistics & delivery
            block; null when the seller provided none. Older clients may ignore it.
        contact_token:
          type: string
          nullable: true
          description: >
            Single-use UUID token minted server-side when an authenticated
            non-owner loads this listing. Required by POST /v1/leads/contact.
            Null for the listing owner and for unauthenticated viewers.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: >
            True when the seller opted this listing in to WhatsApp contact.
            Opt-in only (default false); clients gate the WhatsApp CTA on it.
...
    AvailabilityRange:
      type: object
      description: A booked (unavailable) date range for a furnished/daily rental.
      required: [check_in, check_out]
      properties:
        check_in:
          type: string
        check_out:
          type: string
...
    CreateBookingBody:
      type: object
      required: [check_in, check_out]
      properties:
        check_in:
          type: string
          description: YYYY-MM-DD
        check_out:
          type: string
          description: YYYY-MM-DD
        guests:
          type: integer
        note:
          type: string
          nullable: true
...
paths:
...
  /v1/listings:
    get:
      operationId: getListings
      summary: Browse public listings
      tags: [listings]
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: category
          schema:
            type: string
            enum: [car, real_estate, industrial]
      responses:
        "200":
          description: Paginated listing feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    post:
      operationId: createListing
      summary: Create listing
      tags: [listings]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, category, location, specs, media]
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                  enum: [car, real_estate, industrial]
                base_price_cash:
                  type: number
                  description: Required for normal sale listings. Optional when is_request is true (a buyer request).
                is_request:
                  type: boolean
                  description: When true, this is a buyer "request/wanted" post. Price (base_price_cash) becomes optional and a description is required.
                location:
                  type: string
                latitude:
                  type: number
                  description: Optional precise pin (send with longitude). Overrides the area centroid for near-me search + map display.
                longitude:
                  type: number
                  description: Optional precise pin (send with latitude).
                specs:
                  type: object
                  additionalProperties: true
                media:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [image, video]
                      url:
                        type: string
                      thumbnail_url:
                        type: string
                      is_thumbnail:
                        type: boolean
                payment_options:
                  type: array
                  items:
                    type: object
                    properties:
                      mode:
                        type: string
                        enum: [cash, seller_installment, bank_finance]
                      down_payment:
                        type: number
                      monthly_payment:
                        type: number
                      duration_months:
                        type: integer
                      is_islamic_compliant:
                        type: boolean
                      # P8/M8: declared murabaha/interest rate (0-100). Engine
                      # input only — public offers never expose it.
                      profit_rate_pct:
                        type: number
                logistics:
                  type: object
                  description: >
                    Additive (Task #40). Optional seller-provided logistics &
                    delivery block. All fields optional/nullable.
                  properties:
                    delivery_time_days:
                      type: integer
                      nullable: true
                    origin_type:
                      type: string
                      enum: [local, imported]
                      nullable: true
                    country_of_origin:
                      type: string
                      nullable: true
                    shipping_method:
                      type: string
                      enum: [container, bulk, air]
                      nullable: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/listings/{id}:
    get:
      operationId: getListing
      summary: Get listing detail
      description: >
        Returns full listing detail for active listings. Non-active listings
        (sold, archived) are only visible to their owner; all other callers
        receive 404, preventing access to withdrawn inventory and seller
        contact details.
      tags: [listings]
      security:
        - {}
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ListingDetail"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
        "404":
          description: Not found or listing is non-active and caller is not the owner
    patch:
      operationId: updateListing
      summary: Update own listing (partial)
      tags: [listings]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                base_price_cash:
                  type: number
                location:
                  type: string
                status:
                  type: string
                  enum: [active, sold, archived]
                  description: >
                    Lifecycle status. Sellers set "sold" to mark the deal closed
                    or "archived" to hide the listing.
                specs:
                  type: object
                  additionalProperties: true
                logistics:
                  type: object
                  description: >
                    Additive (Task #40). Optional logistics & delivery patch. All
                    fields optional/nullable.
                  properties:
                    delivery_time_days:
                      type: integer
                      nullable: true
                    origin_type:
                      type: string
                      enum: [local, imported]
                      nullable: true
                    country_of_origin:
                      type: string
                      nullable: true
                    shipping_method:
                      type: string
                      enum: [container, bulk, air]
                      nullable: true
                media:
                  type: array
                  description: >
                    Replace listing media in seller order. Omit to leave photos
                    unchanged. Sale listings must keep at least one item.
                  items:
                    type: object
                    required: [type, url]
                    properties:
                      type:
                        type: string
                        enum: [image, video]
                      url:
                        type: string
                        format: uri
                      thumbnail_url:
                        type: string
                        format: uri
                      is_thumbnail:
                        type: boolean
                      width:
                        type: integer
                      height:
                        type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      updated:
                        type: boolean
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
        "404":
          description: Not found
    delete:
      operationId: deleteListing
      summary: Delete own listing
      tags: [listings]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      deleted:
                        type: boolean
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
        "404":
          description: Not found
  /v1/listings/{id}/similar:
    get:
      operationId: getSimilarListings
      summary: Get similar listings
      tags: [listings]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Similar listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/listings/{id}/availability:
    get:
      operationId: getListingAvailability
      summary: Booked date ranges for a furnished/daily rental
      description: >-
        The date ranges already reserved for a furnished/daily (hotel‑model)
        rental, so the client can grey them out on the calendar. Empty for
        non‑daily listings.
      tags: [listings]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booked ranges
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AvailabilityRange"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/listings/{id}/bookings:
    post:
      operationId: createBooking
      summary: Request a short‑stay booking (furnished/daily rental only)
      description: >-
        Reserve a furnished/daily real‑estate listing for a date range (hotel
        model). Allowed ONLY when the listing's rental_term is furnished_daily;
        long‑term rent and sale are not bookable. Prevents double‑booking. No
        payment yet — this is a request/hold.
      tags: [listings]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "200":
          description: Booking created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Booking"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/listings/{id}/bump:
    post:
      operationId: bumpListing
      summary: Recycle (renew) a listing to the top of recency feeds
      description: >
        Owner-only. Sets bumped_at to now so the listing sorts by
        COALESCE(bumped_at, created_at) in recency feeds and search. NEVER
        changes created_at — the true publish date is preserved. Rate-limited
        with a cooldown; only active, publicly visible listings can be recycled.
      tags: [listings]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing recycled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      bumped_at:
                        type: string
                        description: ISO 8601 timestamp the listing was recycled to.
                      next_bump_available_at:
                        type: string
                        description: ISO 8601 timestamp when the listing may be recycled again (cooldown end).
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
        "403":
          description: Caller is not the listing owner
        "404":
          description: Listing not found, or not eligible to recycle (inactive/hidden)
        "429":
          description: Cooldown has not elapsed (rate limited)
...